NVIDIA: SAUCE: Add support for custom ARM FFH offset handler
The FFH (Functional Fixed Hardware) operation region is maintained by
ARM in https://developer.arm.com/documentation/den0048/latest/

OperationRegion (RegionName, RegionSpace, Offset, Length)

For ARM FFH, Offset is used to identify the functionality offered by
this FFH address space. It must be set to one of the following values:

- 0x0 to indicate usage of 32-bit calling convention
- 0x1 to indicate usage of 64-bit calling convention.
- All other values are reserved.

For GB10 and other similar SOC’s, to communicate with embedded controller,
a new specification is being defined. It is currently in draft stage and
maintained in

https://github.com/OpenDevicePartnership/documentation/blob/main/bookshelf/Shelf%204%20Specifications/EC%20Interface/src/README.md
https://github.com/OpenDevicePartnership/documentation/blob/main/bookshelf/Shelf%204%20Specifications/EC%20Interface/src/secure-ec-services-overview.md

Offset 4 section:

https://github.com/OpenDevicePartnership/documentation/blob/main/bookshelf/Shelf%204%20Specifications/EC%20Interface/src/secure-ec-services-overview.md#operation-region-definition

This specification internally uses offset 0x4 which is not defined in
published ARM specification. So, when ACPI request comes with offset 0x4,
then it will fail due to missing support. This commit adds support for
custom offset handler. A new EC interface driver will be added in
subsequent patches which will registers it callback function.
When FFH operation region will be executed with offsets other
than 0x0 and 0x1, then it will be forwarded to custom handler.

Signed-off-by: Abhishek Sahu <abhsahu@nvidia.com>